refactor(protocol): split the negotiation graph into top-level nodes
